+++ documentation/release-notes.markdown
@@ -7,9 +7,11 @@
   * Always apply the MIME type of a file when uploading to S3.
   * Speed up previewing a site with dynamic generation by only re-reading
     source files that have changed, instead of everything each time.
-  * Add a [generator for compiling SASS][sass] into CSS
+  * Add a [generator for compiling SASS][sass] into CSS.
+  * Add [generators][cal] for calendar-based index pages.
 
 [sass]: https://flourish.readthedocs.io/en/latest/api-flourish-generators-sass/
+[cal]: https://flourish.readthedocs.io/en/latest/api-flourish-generators-calendar/
 
 ## 0.7 — 16 June 2020